+++ packages/jsenv-plugin-preact/tests/preact_and_redux/preact_and_redux_build.test.mjs
@@ -19,7 +19,7 @@ const test = async (params) => {
     },
     buildDirectoryUrl: new URL("./dist/", import.meta.url),
     plugins: [...plugins, jsenvPluginBundling()],
-    writeGeneratedFiles: true,
+    outDirectoryUrl: new URL("./.jsenv/", import.meta.url),
     ...params,
   })
   const server = await startFileServer({