+++ packages/jsenv-plugin-preact/tests/script_module_jsx_error/script_module_jsx_error_dev.test.mjs
@@ -4,7 +4,7 @@ import { startDevServer } from "@jsenv/core"
 
 const devServer = await startDevServer({
   logLevel: "warn",
-  rootDirectoryUrl: new URL("./client/", import.meta.url),
+  sourceDirectoryUrl: new URL("./client/", import.meta.url),
   keepProcessAlive: false,
 })
 const browser = await chromium.launch({ headless: true })