+++ packages/pwa/tests/start_test_server.mjs
@@ -6,7 +6,7 @@ export const startTestServer = async (rest) => {
   const testServer = await startDevServer({
     logLevel: "warn",
     https: { certificate, privateKey },
-    rootDirectoryUrl: new URL("./client/", import.meta.url),
+    sourceDirectoryUrl: new URL("./client/", import.meta.url),
     clientAutoreload: false,
     supervisor: false,
     ...rest,