+++ src/plugins/transpilation/as_js_classic/jsenv_plugin_as_js_classic_library.js
@@ -49,7 +49,7 @@ export const jsenvPluginAsJsClassicLibrary = ({
         jsModuleUrlInfos: [jsModuleUrlInfo],
         context: {
           ...context,
-          buildDirectoryUrl: context.outDirectoryUrl,
+          buildDirectoryUrl: new URL("./", import.meta.url),
         },
         preserveDynamicImport: true,
       })