+++ tests/dev_and_build/import_assert_type_css/import_assert_type_css_build.test.mjs
@@ -18,7 +18,7 @@ const test = async ({ snapshotsDirectoryName, ...rest }) => {
       "./main.html": "main.html",
     },
     buildDirectoryUrl: new URL("./dist/", import.meta.url),
-    writeGeneratedFiles: true,
+    outDirectoryUrl: new URL("./.jsenv/", import.meta.url),
     ...rest,
   })
   const server = await startFileServer({