+++ tests/dev_and_build/top_level_await_2/top_level_await_2.test.mjs
@@ -12,7 +12,7 @@ const test = async (params) => {
       "./main.html": "main.html",
     },
     buildDirectoryUrl: new URL("./dist/", import.meta.url),
-    writeGeneratedFiles: true,
+    outDirectoryUrl: new URL("./.jsenv/", import.meta.url),
     ...params,
   })