+++ tests/dev_and_build/workers/worker_versioning_importmap/worker_versioning_importmap_build.test.mjs
@@ -21,7 +21,7 @@ const test = async ({ snapshotsDirectoryUrl, ...rest }) => {
       "./main.html": "main.html",
     },
     buildDirectoryUrl: new URL("./dist/", import.meta.url),
-    writeGeneratedFiles: true,
+    outDirectoryUrl: new URL("./.jsenv/", import.meta.url),
     ...rest,
   })